fixed no-ball ball count
