+++ .vscode/settings.json
@@ -1,4 +1,8 @@
 {
     "editor.formatOnSave": true,
     "python.formatting.provider": "black",
+    "python.formatting.blackArgs": [
+        "--line-length",
+        "119"
+    ],
 }
