--- simulations/06d90aabbed24a87a9fc8dd9ab0d1c01/first_innings_batting.txt
@@ -1,12 +1,0 @@
-Yashasvi Jaiswal     		 c & b Avesh Khan                         		 67  (25)
-Jos Buttler          		 b Ankit Rajpoot                          		 14  (7)
-Sanju Samson         		 c Quinton de Kock b Ravi Bishnoi         		 1   (6)
-Rassie van der Dussen 		 c Ravi Bishnoi b Mark Wood               		 2   (3)
-Karun Nair           		 b Mark Wood                              		 6   (4)
-Jimmy Neesham        		 c & b Marcus Stoinis                     		 74  (26)
-Riyan Parag          		 b Marcus Stoinis                         		 47  (14)
-Ravichandran Ashwin  		 c Quinton de Kock b Marcus Stoinis       		 0   (1)
-Trent Boult          		 not out                                  		 62  (19)
-Prasidh Krishna      		 c Deepak Hooda b Ravi Bishnoi            		 31  (10)
-Yuzvendra Chahal     		 lbw b Ravi Bishnoi                       		 4   (2)
-extras               		                                          		 15
