--- simulations/06d90aabbed24a87a9fc8dd9ab0d1c01/first_innings_bowling.txt
@@ -1,7 +1,0 @@
-Deepak Hooda         	 3 - 62 - 0
-Mark Wood            	 3 - 44 - 2
-Ravi Bishnoi         	 3.3 - 41 - 3
-Marcus Stoinis       	 4 - 42 - 3
-Ankit Rajpoot        	 1 - 13 - 1
-Avesh Khan           	 3 - 60 - 1
-Shahbaz Nadeem       	 2 - 46 - 0
