--- simulations/06d90aabbed24a87a9fc8dd9ab0d1c01/first_innings_fow.txt
@@ -1,10 +1,0 @@
-Jos Buttler          65/1
-Sanju Samson         72/2
-Rassie van der Dussen 76/3
-Karun Nair           82/4
-Yashasvi Jaiswal     92/5
-Riyan Parag          196/6
-Ravichandran Ashwin  196/7
-Jimmy Neesham        245/8
-Prasidh Krishna      319/9
-Yuzvendra Chahal     323/10
