--- simulations/06d90aabbed24a87a9fc8dd9ab0d1c01/second_innings_batting.txt
@@ -1,12 +1,0 @@
-Quinton de Kock      		 b Trent Boult                            		 7   (4)
-Evin Lewis           		 lbw b Yuzvendra Chahal                   		 0   (1)
-Lokesh Rahul         		 lbw b Riyan Parag                        		 11  (8)
-Marcus Stoinis       		 c Riyan Parag b Trent Boult              		 0   (1)
-Manish Pandey        		 c Sanju Samson b Trent Boult             		 0   (1)
-Deepak Hooda         		 b Trent Boult                            		 7   (7)
-Mark Wood            		 lbw b Ravichandran Ashwin                		 10  (7)
-Shahbaz Nadeem       		 lbw b Trent Boult                        		 0   (1)
-Avesh Khan           		 c Jos Buttler b Prasidh Krishna          		 63  (27)
-Ankit Rajpoot        		 b Jimmy Neesham                          		 55  (24)
-Ravi Bishnoi         		 not out                                  		 27  (10)
-extras               		                                          		 9
