--- simulations/06d90aabbed24a87a9fc8dd9ab0d1c01/second_innings_bowling.txt
@@ -1,6 +1,0 @@
-Yuzvendra Chahal     	 2 - 29 - 1
-Trent Boult          	 3 - 22 - 5
-Riyan Parag          	 4 - 49 - 1
-Ravichandran Ashwin  	 2 - 22 - 1
-Prasidh Krishna      	 3 - 46 - 1
-Jimmy Neesham        	 1.1 - 12 - 1
