--- simulations/06d90aabbed24a87a9fc8dd9ab0d1c01/second_innings_fow.txt
@@ -1,10 +1,0 @@
-Evin Lewis           7/1
-Quinton de Kock      16/2
-Marcus Stoinis       16/3
-Manish Pandey        16/4
-Lokesh Rahul         21/5
-Deepak Hooda         25/6
-Shahbaz Nadeem       25/7
-Mark Wood            50/8
-Avesh Khan           138/9
-Ankit Rajpoot        189/10
