--- simulations/06d90aabbed24a87a9fc8dd9ab0d1c01/second_innings_over_by_over.txt
@@ -1,16 +1,0 @@
-Over: 1, Bowler: Yuzvendra Chahal,['4', '2', '1', 'W', '6', '3'], 16/1
-Over: 2, Bowler: Trent Boult,['W', 'W', 'W', '0', '2', '1'], 3/3
-Over: 3, Bowler: Riyan Parag,['2', '0', '0', '0', '0', 'W'], 2/1
-Over: 4, Bowler: Trent Boult,['4', '0', '0', 'W', 'W', '1'], 5/2
-Over: 5, Bowler: Riyan Parag,['1', '2', '3', '4', '0', '3'], 13/0
-Over: 6, Bowler: Ravichandran Ashwin,['2', '6', '1', '2', '0', 'W'], 11/1
-Over: 7, Bowler: Prasidh Krishna,['2', '2', '2', '0', 'Nb', '0', '4'], 10/0
-Over: 8, Bowler: Jimmy Neesham,['4', '0', '1', '1', 'Nb', '6', '0'], 12/0
-Over: 9, Bowler: Riyan Parag,['2', '4', 'Nb', '3', '6', 'Nb', '3', '3'], 21/0
-Over: 10, Bowler: Trent Boult,['Nb', '2', '1', '6', '1', '1', '3'], 14/0
-Over: 11, Bowler: Yuzvendra Chahal,['1', '2', '2', '4', '4', '0'], 13/0
-Over: 12, Bowler: Prasidh Krishna,['1', '3', '4', '2', '3', 'W'], 13/1
-Over: 13, Bowler: Riyan Parag,['2', '2', '2', '4', '0', '3'], 13/0
-Over: 14, Bowler: Ravichandran Ashwin,['Nb', '2', '1', '3', '0', '1', '4'], 11/0
-Over: 15, Bowler: Prasidh Krishna,['2', '6', '2', '6', 'Wd', 'Nb', '4', 'Nb', '3'], 23/0
-Over: 16, Bowler: Jimmy Neesham,['W'], 0/1
